refactor(ghostmap): split Re-Fly suppression cache into GhostMapPresence.ReFlySuppressionCache.cs
